added display and write functions, working correctly, fixed some bugs in argument parsing and updated formatting in a few other functions

+++ Project1/include/io.h
@@ -32,12 +32,12 @@ struct io cmd_get();
 
 void io_parse(char* arg_string, uint8_t arg_qty, ...);
 
-uint32_t string_num(char* string);
+uint64_t string_num(char* string);
 
-uint32_t string_dec(char* string);
+uint64_t string_dec(char* string);
 
-uint32_t string_hex(char* string);
+uint64_t string_hex(char* string);
 
-uint32_t power(uint32_t base, uint32_t exponent);
+uint64_t power(uint32_t base, uint32_t exponent);
 
 #endif //IO_H
